+++ javascript/WebGL/createFrameRenderer.js
@@ -9,25 +9,31 @@ import forEach from 'lodash/forEach'
 const { requestAnimationFrame } = window
 
 let time = Date.now()
-export default ({ canvas, context, shader, program, store, startTime = Date.now() }) => function renderFrame () {
+export default ({ canvas, context, shader, program, store }) => function renderFrame () {
   /* eslint-disable no-multi-spaces, key-spacing */
   const state = store.getState()
 
   const currentShader = getCurrentShader(state)
   if (shader === currentShader) {
-    const { center, range } = getShaderViewport(state, currentShader)
+    const { center, range, rotation } = getShaderViewport(state, currentShader)
     const config            = getShaderConfig(state, currentShader)
 
     if (config.speed) {
       time += parseFloat(config.speed)
     } else {
+      /* only here for spinning cube */
       time += 0.016
     }
+
+    const ASPECT_RATIO = window.innerWidth / window.innerHeight
 
     const uniformValues = assign({}, config, {
       shader: SHADER_ENUM[currentShader],
       center: [center.x, center.y],
-      range:  [range.x, range.y],
+      /* range.x is intentionally ignored in favor of setting */
+      /* the window dimensions to dictate aspect ratio */
+      range:  [range.y * ASPECT_RATIO, range.y],
+      rotation: rotation || 0,
       resolution: [
         window.innerWidth,
         window.innerHeight
@@ -37,7 +43,8 @@ export default ({ canvas, context, shader, program, store, startTime = Date.now(
         0.2321 + Math.cos(time / 1330) / 40
       ],
       msaa_coordinates: msaaCoordinates[config.supersamples],
-      time: time - startTime // large times won't convert to float 32 well
+      /* large times won't convert to float 32 well :( */
+      time: time
     })
 
     context.useProgram(program)