+++ javascript/reducers/index.js
@@ -1,5 +1,6 @@
 import { combineReducers } from 'redux'
 import { DEFAULT_STORE } from 'javascript/config'
+import createReducer from 'reducers/createReducer'
 import currentShader, * as fromCurrentShader from 'reducers/currentShader'
 import createShader, * as Shader from 'reducers/createShader'
 import menuOpen from 'reducers/menuOpen'
@@ -7,12 +8,14 @@ import mapValues from 'lodash/mapValues'
 
 export default combineReducers({
   currentShader,
+  pinchStart: createReducer('pinch_start', {}),
+  menuOpen,
   shaders: combineReducers(mapValues(DEFAULT_STORE, (shaderConfig, shaderName) => (
     createShader(shaderName, shaderConfig)
-  ))),
-  menuOpen
+  )))
 })
 
 export const getCurrentShader = state => fromCurrentShader.getCurrentShader(state)
 export const getShaderConfig = (state, shader) => Shader.getShaderConfig(state, shader)
 export const getShaderViewport = (state, shader) => Shader.getShaderViewport(state, shader)
+export const getPinchStart = state => state.pinchStart